Move create_and_store_nonrevocable_credential to the only place of usage
