+++ aries_vcx/src/common/proofs/verifier/verifier.rs
@@ -52,13 +52,261 @@ pub async fn validate_indy_proof(
 
 #[cfg(test)]
 #[allow(clippy::unwrap_used)]
-pub mod unit_tests {
-    use crate::common::proofs::proof_request::ProofRequestData;
-    use crate::common::test_utils::create_and_store_nonrevocable_credential;
-    use crate::utils;
-    use crate::utils::devsetup::SetupProfile;
-
+pub mod integration_tests {
     use super::*;
+
+    use std::sync::Arc;
+
+    use aries_vcx_core::{
+        anoncreds::base_anoncreds::BaseAnonCreds,
+        ledger::base_ledger::{AnoncredsLedgerRead, AnoncredsLedgerWrite},
+    };
+
+    use crate::{
+        common::{
+            credentials::encoding::encode_attributes,
+            proofs::proof_request::ProofRequestData,
+            test_utils::{create_and_write_test_cred_def, create_and_write_test_schema_1},
+        },
+        errors::error::AriesVcxErrorKind,
+        global::settings,
+        utils::{self, constants::DEFAULT_SCHEMA_ATTRS, devsetup::SetupProfile},
+    };
+
+    // FUTURE - issuer and holder seperation only needed whilst modular deps not fully implemented
+    async fn create_indy_proof(
+        anoncreds_issuer: &Arc<dyn BaseAnonCreds>,
+        anoncreds_holder: &Arc<dyn BaseAnonCreds>,
+        ledger_read: &Arc<dyn AnoncredsLedgerRead>,
+        ledger_write: &Arc<dyn AnoncredsLedgerWrite>,
+        did: &str,
+    ) -> (String, String, String, String) {
+        let (schema_id, schema_json, cred_def_id, cred_def_json, _offer, _req, _req_meta, cred_id) =
+            create_and_store_nonrevocable_credential(
+                anoncreds_issuer,
+                anoncreds_holder,
+                ledger_read,
+                ledger_write,
+                &did,
+                DEFAULT_SCHEMA_ATTRS,
+            )
+            .await;
+        let proof_req = json!({
+           "nonce":"123432421212",
+           "name":"proof_req_1",
+           "version":"0.1",
+           "requested_attributes": json!({
+               "address1_1": json!({
+                   "name":"address1",
+                   "restrictions": [json!({ "issuer_did": did })]
+               }),
+               "zip_2": json!({
+                   "name":"zip",
+                   "restrictions": [json!({ "issuer_did": did })]
+               }),
+               "self_attest_3": json!({
+                   "name":"self_attest",
+               }),
+           }),
+           "requested_predicates": json!({}),
+        })
+        .to_string();
+        let requested_credentials_json = json!({
+              "self_attested_attributes":{
+                 "self_attest_3": "my_self_attested_val"
+              },
+              "requested_attributes":{
+                 "address1_1": {"cred_id": cred_id, "revealed": true},
+                 "zip_2": {"cred_id": cred_id, "revealed": true}
+                },
+              "requested_predicates":{}
+        })
+        .to_string();
+
+        let schema_json: serde_json::Value = serde_json::from_str(&schema_json).unwrap();
+        let schemas = json!({
+            schema_id: schema_json,
+        })
+        .to_string();
+
+        let cred_def_json: serde_json::Value = serde_json::from_str(&cred_def_json).unwrap();
+        let cred_defs = json!({
+            cred_def_id: cred_def_json,
+        })
+        .to_string();
+
+        anoncreds_holder
+            .prover_get_credentials_for_proof_req(&proof_req)
+            .await
+            .unwrap();
+
+        let proof = anoncreds_holder
+            .prover_create_proof(
+                &proof_req,
+                &requested_credentials_json,
+                "main",
+                &schemas,
+                &cred_defs,
+                None,
+            )
+            .await
+            .unwrap();
+        (schemas, cred_defs, proof_req, proof)
+    }
+
+    async fn create_proof_with_predicate(
+        anoncreds_issuer: &Arc<dyn BaseAnonCreds>,
+        anoncreds_holder: &Arc<dyn BaseAnonCreds>,
+        ledger_read: &Arc<dyn AnoncredsLedgerRead>,
+        ledger_write: &Arc<dyn AnoncredsLedgerWrite>,
+        did: &str,
+        include_predicate_cred: bool,
+    ) -> (String, String, String, String) {
+        let (schema_id, schema_json, cred_def_id, cred_def_json, _offer, _req, _req_meta, cred_id) =
+            create_and_store_nonrevocable_credential(
+                anoncreds_issuer,
+                anoncreds_holder,
+                ledger_read,
+                ledger_write,
+                &did,
+                DEFAULT_SCHEMA_ATTRS,
+            )
+            .await;
+
+        let proof_req = json!({
+           "nonce":"123432421212",
+           "name":"proof_req_1",
+           "version":"0.1",
+           "requested_attributes": json!({
+               "address1_1": json!({
+                   "name":"address1",
+                   "restrictions": [json!({ "issuer_did": did })]
+               }),
+               "self_attest_3": json!({
+                   "name":"self_attest",
+               }),
+           }),
+           "requested_predicates": json!({
+               "zip_3": {"name":"zip", "p_type":">=", "p_value":18}
+           }),
+        })
+        .to_string();
+
+        let requested_credentials_json;
+        if include_predicate_cred {
+            requested_credentials_json = json!({
+              "self_attested_attributes":{
+                 "self_attest_3": "my_self_attested_val"
+              },
+              "requested_attributes":{
+                 "address1_1": {"cred_id": cred_id, "revealed": true}
+                },
+              "requested_predicates":{
+                  "zip_3": {"cred_id": cred_id}
+              }
+            })
+            .to_string();
+        } else {
+            requested_credentials_json = json!({
+              "self_attested_attributes":{
+                 "self_attest_3": "my_self_attested_val"
+              },
+              "requested_attributes":{
+                 "address1_1": {"cred_id": cred_id, "revealed": true}
+                },
+              "requested_predicates":{
+              }
+            })
+            .to_string();
+        }
+
+        let schema_json: serde_json::Value = serde_json::from_str(&schema_json).unwrap();
+        let schemas = json!({
+            schema_id: schema_json,
+        })
+        .to_string();
+
+        let cred_def_json: serde_json::Value = serde_json::from_str(&cred_def_json).unwrap();
+        let cred_defs = json!({
+            cred_def_id: cred_def_json,
+        })
+        .to_string();
+
+        anoncreds_holder
+            .prover_get_credentials_for_proof_req(&proof_req)
+            .await
+            .unwrap();
+
+        let proof = anoncreds_holder
+            .prover_create_proof(
+                &proof_req,
+                &requested_credentials_json,
+                "main",
+                &schemas,
+                &cred_defs,
+                None,
+            )
+            .await
+            .unwrap();
+        (schemas, cred_defs, proof_req, proof)
+    }
+
+    async fn create_and_store_nonrevocable_credential(
+        anoncreds_issuer: &Arc<dyn BaseAnonCreds>,
+        anoncreds_holder: &Arc<dyn BaseAnonCreds>,
+        ledger_read: &Arc<dyn AnoncredsLedgerRead>,
+        ledger_write: &Arc<dyn AnoncredsLedgerWrite>,
+        issuer_did: &str,
+        attr_list: &str,
+    ) -> (String, String, String, String, String, String, String, String) {
+        let schema = create_and_write_test_schema_1(anoncreds_issuer, ledger_write, issuer_did, attr_list).await;
+
+        let cred_def = create_and_write_test_cred_def(
+            anoncreds_issuer,
+            ledger_read,
+            ledger_write,
+            issuer_did,
+            &schema.schema_id,
+        )
+        .await;
+
+        let cred_def_id = cred_def.get_cred_def_id();
+        let cred_def_json = ledger_read.get_cred_def(&cred_def_id, Some(issuer_did)).await.unwrap();
+
+        let offer = anoncreds_issuer
+            .issuer_create_credential_offer(&cred_def_id)
+            .await
+            .unwrap();
+        let master_secret_name = settings::DEFAULT_LINK_SECRET_ALIAS;
+        let (req, req_meta) = anoncreds_holder
+            .prover_create_credential_req(&issuer_did, &offer, &cred_def_json, master_secret_name)
+            .await
+            .unwrap();
+
+        /* create cred */
+        let credential_data = r#"{"address1": ["123 Main St"], "address2": ["Suite 3"], "city": ["Draper"], "state": ["UT"], "zip": ["84000"]}"#;
+        let encoded_attributes = encode_attributes(&credential_data).unwrap();
+
+        let (cred, _, _) = anoncreds_issuer
+            .issuer_create_credential(&offer, &req, &encoded_attributes, None, None)
+            .await
+            .unwrap();
+        /* store cred */
+        let cred_id = anoncreds_holder
+            .prover_store_credential(None, &req_meta, &cred, &cred_def_json, None)
+            .await
+            .unwrap();
+        (
+            schema.schema_id,
+            schema.schema_json,
+            cred_def_id,
+            cred_def_json,
+            offer,
+            req,
+            req_meta,
+            cred_id,
+        )
+    }
 
     #[tokio::test]
     #[ignore]
@@ -347,201 +595,6 @@ pub mod unit_tests {
         })
         .await;
     }
-}
-
-#[cfg(test)]
-#[allow(clippy::unwrap_used)]
-pub mod integration_tests {
-    use std::sync::Arc;
-
-    use aries_vcx_core::{
-        anoncreds::base_anoncreds::BaseAnonCreds,
-        ledger::base_ledger::{AnoncredsLedgerRead, AnoncredsLedgerWrite},
-    };
-
-    use crate::{
-        common::test_utils::create_and_store_nonrevocable_credential,
-        utils::{constants::DEFAULT_SCHEMA_ATTRS, devsetup::SetupProfile},
-    };
-
-    // FUTURE - issuer and holder seperation only needed whilst modular deps not fully implemented
-    async fn create_indy_proof(
-        anoncreds_issuer: &Arc<dyn BaseAnonCreds>,
-        anoncreds_holder: &Arc<dyn BaseAnonCreds>,
-        ledger_read: &Arc<dyn AnoncredsLedgerRead>,
-        ledger_write: &Arc<dyn AnoncredsLedgerWrite>,
-        did: &str,
-    ) -> (String, String, String, String) {
-        let (schema_id, schema_json, cred_def_id, cred_def_json, _offer, _req, _req_meta, cred_id) =
-            create_and_store_nonrevocable_credential(
-                anoncreds_issuer,
-                anoncreds_holder,
-                ledger_read,
-                ledger_write,
-                &did,
-                DEFAULT_SCHEMA_ATTRS,
-            )
-            .await;
-        let proof_req = json!({
-           "nonce":"123432421212",
-           "name":"proof_req_1",
-           "version":"0.1",
-           "requested_attributes": json!({
-               "address1_1": json!({
-                   "name":"address1",
-                   "restrictions": [json!({ "issuer_did": did })]
-               }),
-               "zip_2": json!({
-                   "name":"zip",
-                   "restrictions": [json!({ "issuer_did": did })]
-               }),
-               "self_attest_3": json!({
-                   "name":"self_attest",
-               }),
-           }),
-           "requested_predicates": json!({}),
-        })
-        .to_string();
-        let requested_credentials_json = json!({
-              "self_attested_attributes":{
-                 "self_attest_3": "my_self_attested_val"
-              },
-              "requested_attributes":{
-                 "address1_1": {"cred_id": cred_id, "revealed": true},
-                 "zip_2": {"cred_id": cred_id, "revealed": true}
-                },
-              "requested_predicates":{}
-        })
-        .to_string();
-
-        let schema_json: serde_json::Value = serde_json::from_str(&schema_json).unwrap();
-        let schemas = json!({
-            schema_id: schema_json,
-        })
-        .to_string();
-
-        let cred_def_json: serde_json::Value = serde_json::from_str(&cred_def_json).unwrap();
-        let cred_defs = json!({
-            cred_def_id: cred_def_json,
-        })
-        .to_string();
-
-        anoncreds_holder
-            .prover_get_credentials_for_proof_req(&proof_req)
-            .await
-            .unwrap();
-
-        let proof = anoncreds_holder
-            .prover_create_proof(
-                &proof_req,
-                &requested_credentials_json,
-                "main",
-                &schemas,
-                &cred_defs,
-                None,
-            )
-            .await
-            .unwrap();
-        (schemas, cred_defs, proof_req, proof)
-    }
-
-    async fn create_proof_with_predicate(
-        anoncreds_issuer: &Arc<dyn BaseAnonCreds>,
-        anoncreds_holder: &Arc<dyn BaseAnonCreds>,
-        ledger_read: &Arc<dyn AnoncredsLedgerRead>,
-        ledger_write: &Arc<dyn AnoncredsLedgerWrite>,
-        did: &str,
-        include_predicate_cred: bool,
-    ) -> (String, String, String, String) {
-        let (schema_id, schema_json, cred_def_id, cred_def_json, _offer, _req, _req_meta, cred_id) =
-            create_and_store_nonrevocable_credential(
-                anoncreds_issuer,
-                anoncreds_holder,
-                ledger_read,
-                ledger_write,
-                &did,
-                DEFAULT_SCHEMA_ATTRS,
-            )
-            .await;
-
-        let proof_req = json!({
-           "nonce":"123432421212",
-           "name":"proof_req_1",
-           "version":"0.1",
-           "requested_attributes": json!({
-               "address1_1": json!({
-                   "name":"address1",
-                   "restrictions": [json!({ "issuer_did": did })]
-               }),
-               "self_attest_3": json!({
-                   "name":"self_attest",
-               }),
-           }),
-           "requested_predicates": json!({
-               "zip_3": {"name":"zip", "p_type":">=", "p_value":18}
-           }),
-        })
-        .to_string();
-
-        let requested_credentials_json;
-        if include_predicate_cred {
-            requested_credentials_json = json!({
-              "self_attested_attributes":{
-                 "self_attest_3": "my_self_attested_val"
-              },
-              "requested_attributes":{
-                 "address1_1": {"cred_id": cred_id, "revealed": true}
-                },
-              "requested_predicates":{
-                  "zip_3": {"cred_id": cred_id}
-              }
-            })
-            .to_string();
-        } else {
-            requested_credentials_json = json!({
-              "self_attested_attributes":{
-                 "self_attest_3": "my_self_attested_val"
-              },
-              "requested_attributes":{
-                 "address1_1": {"cred_id": cred_id, "revealed": true}
-                },
-              "requested_predicates":{
-              }
-            })
-            .to_string();
-        }
-
-        let schema_json: serde_json::Value = serde_json::from_str(&schema_json).unwrap();
-        let schemas = json!({
-            schema_id: schema_json,
-        })
-        .to_string();
-
-        let cred_def_json: serde_json::Value = serde_json::from_str(&cred_def_json).unwrap();
-        let cred_defs = json!({
-            cred_def_id: cred_def_json,
-        })
-        .to_string();
-
-        anoncreds_holder
-            .prover_get_credentials_for_proof_req(&proof_req)
-            .await
-            .unwrap();
-
-        let proof = anoncreds_holder
-            .prover_create_proof(
-                &proof_req,
-                &requested_credentials_json,
-                "main",
-                &schemas,
-                &cred_defs,
-                None,
-            )
-            .await
-            .unwrap();
-        (schemas, cred_defs, proof_req, proof)
-    }
-
     #[tokio::test]
     #[ignore]
     async fn test_pool_prover_verify_proof() {
